test: restructure test_backend_failure test suite
break test_backend_failure tests into several files to improve the tests
overall running time
